requirement.txt
fastapi
uvicorn[standard ]
sqlalchemy
bcrypt==3.2.0
email-validator==1.1.2
pydantic
PyJWT==2.0.1
passlib